add opt example

+++ examples/opt/opt_server.py
@@ -8,11 +8,13 @@ from energonai.engine import InferenceEngine
 
 from transformers import GPT2Tokenizer
 
-app = FastAPI() # 创建 api 对象
+app = FastAPI()  # 创建 api 对象
 
-@app.get("/") # 根路由
+
+@app.get("/")  # 根路由
 def root():
     return {"200"}
+
 
 @app.get("/run/{request}")
 def run(request: str, max_seq_length: int):
@@ -28,9 +30,9 @@ def run(request: str, max_seq_length: int):
         if '<|endoftext|>' in total_predicted_text:
             break
         input_token = tokenizer(total_predicted_text, return_tensors="pt")
-    
+
     return {total_predicted_text}
-    
+
 
 @app.get("/shutdown")
 async def shutdown():
@@ -41,40 +43,40 @@ async def shutdown():
 
 
 def launch_engine(model_class,
-                model_type,
-                max_batch_size: int = 1,
-                tp_init_size: int = -1,
-                pp_init_size: int = -1,
-                host: str = "localhost",
-                port: int = 29500,
-                dtype = torch.float,
-                checkpoint: str = None,
-                tokenizer_path: str = None,
-                server_host = "localhost",
-                server_port = 8005,
-                log_level = "info"
-                ):
-    
+                  model_type,
+                  max_batch_size: int = 1,
+                  tp_init_size: int = -1,
+                  pp_init_size: int = -1,
+                  host: str = "localhost",
+                  port: int = 29500,
+                  dtype=torch.float,
+                  checkpoint: str = None,
+                  tokenizer_path: str = None,
+                  server_host="localhost",
+                  server_port=8005,
+                  log_level="info"
+                  ):
+
     # only for the generation task
     global tokenizer
     if(tokenizer_path):
         tokenizer = GPT2Tokenizer.from_pretrained(tokenizer_path)
-    
+
     if checkpoint:
         model_config = {'dtype': dtype, 'checkpoint': checkpoint}
     else:
         model_config = {'dtype': dtype}
-    
+
     global engine
-    engine = InferenceEngine(model_class, 
-                            model_config,
-                            model_type,
-                            max_batch_size = max_batch_size, 
-                            tp_init_size = tp_init_size, 
-                            pp_init_size = pp_init_size, 
-                            host = host,
-                            port = port,
-                            dtype = dtype)
+    engine = InferenceEngine(model_class,
+                             model_config,
+                             model_type,
+                             max_batch_size=max_batch_size,
+                             tp_init_size=tp_init_size,
+                             pp_init_size=pp_init_size,
+                             host=host,
+                             port=port,
+                             dtype=dtype)
 
     global server
     config = uvicorn.Config(app, host=server_host, port=server_port, log_level=log_level)
